- [add] 修改商城店铺接口

+++ api-mall/src/main/java/com/hailu/cloud/api/mall/module/multiindustry/entity/StoreInformation.java
@@ -90,7 +90,7 @@ public class StoreInformation implements Serializable {
      * 营业状态(1-营业中，2-休息中)
      */
     @ApiModelProperty("营业状态(1-营业中，2-休息中)")
-    private Long businessState;
+    private Integer businessState;
 
     /**
      * 营业状态(1-营业中，2-休息中)
@@ -127,7 +127,7 @@ public class StoreInformation implements Serializable {
      * 审核('审核中-1','审核通过-2','审核不通过-3')
      */
     @ApiModelProperty("审核('审核中-1','审核通过-2','审核不通过-3')")
-    private Long toExamine;
+    private Integer toExamine;
 
     @DictName(code = "TO_EXAMINE", joinField = "toExamine")
     @ApiModelProperty("审核('审核中','审核通过','审核不通过')")
